--- a/Code/Framework/AzCore/Tests/azcoretests_files.cmake
+++ b/Code/Framework/AzCore/Tests/azcoretests_files.cmake
@@ -134,6 +134,7 @@ set(FILES
     Math/Matrix4x4PerformanceTests.cpp
     Math/Matrix4x4Tests.cpp
     Math/MatrixMxNTests.cpp
+    Math/MatrixMxNPerformanceTests.cpp
     Math/MatrixUtilsTests.cpp
     Math/MathTest.h
     Math/MathTestData.h
@@ -161,6 +162,7 @@ set(FILES
     Math/Vector4PerformanceTests.cpp
     Math/Vector4Tests.cpp
     Math/VectorNTests.cpp
+    Math/VectorNPerformanceTests.cpp
     Memory/AllocatorBenchmarks.cpp
     Memory/HphaAllocator.cpp
     Memory/HphaAllocatorErrorDetection.cpp