--- a/Gems/ScriptCanvas/Code/Include/ScriptCanvas/Libraries/Math/VectorN.cpp
+++ b/Gems/ScriptCanvas/Code/Include/ScriptCanvas/Libraries/Math/VectorN.cpp
@@ -76,12 +76,12 @@ namespace ScriptCanvas
 
         Data::NumberType LengthSquared(const Data::VectorNType& source)
         {
-            return source.GetLengthSq();
+            return source.Dot(source);
         }
 
         Data::NumberType Length(const Data::VectorNType& source)
         {
-            return source.GetLength();
+            return source.L2Norm();
         }
 
         Data::VectorNType& Normalize(Data::VectorNType& source)